Refactor: PromotionProjectRequestDTO에 잘못넣은거 PromotionProjectResponseDTO로 옮기기 (#57)
