+++ src/main/java/com/partnerd/web/dto/projectDTO/PromotionProjectResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 public class PromotionProjectResponseDTO {
 
@@ -27,5 +28,30 @@ public class PromotionProjectResponseDTO {
     public static class UpdatePromotionProjectResultDTO {
         private Long promotionProjectId;
         private LocalDateTime updatedAt;
+    }
+
+    // 프로젝트 홍보글 모아보기 (한 칸씩)
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class PromotionProjectPreviewDTO {
+        private Long promotionProjectId;
+        private String title;
+        private String intro;
+    }
+
+    // 프로젝트 홍보글 모아보기 (전체 리스트)
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class PromotionProjectPreviewListDTO{
+        private List<PromotionProjectPreviewDTO> promotionProjectPreviewDTOList;
+        private Integer listSize;
+        private Integer totalPage;
+        private Long totalElements;
+        private Boolean isFirst;
+        private Boolean isLast;
     }
 }
